% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{drake-package}
\alias{drake-package}
\alias{drake}
\title{drakeplanner: An R/Shiny app to create and visualize \code{drake} workflows.}
\description{
\code{drakeplanner} is an package and R/Shiny app
for creating and visualizing \code{drake} workflows.
}
\examples{
\dontrun{
drakeplanner()
}
}
\references{
\url{https://github.com/wlandau/drakeplanner}
}
\author{
William Michael Landau \email{will.landau@gmail.com}
}
