% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{drakeplanner}
\alias{drakeplanner}
\title{Launch the drake planner}
\usage{
drakeplanner()
}
\description{
Launch an interactive web application for
creating and visualizing \code{drake} workflows.
}
\examples{
\dontrun{
drakeplanner()
}
}
